# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: ComfyUI Serverless API
  version: 1.0.0
  description: |
    Serverless ComfyUI API for AI image generation and workflow execution.
    
    This API provides a complete solution for running ComfyUI workflows in a serverless environment,
    with features including:
    - Asynchronous workflow execution
    - Model management and caching  
    - File upload/download
    - Real-time progress tracking
    - Webhook notifications
    - Authentication and rate limiting
    
  contact:
    name: API Support
    email: support@comfyui-serverless.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.comfyui-serverless.com
    description: Production server
  - url: https://staging-api.comfyui-serverless.com
    description: Staging server
  - url: http://localhost:8000
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /workflows/execute:
    post:
      summary: Execute Workflow
      description: Submit a ComfyUI workflow for asynchronous execution
      tags:
        - Workflows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowExecutionRequest'
      responses:
        '200':
          description: Workflow submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionResponse'
        '400':
          description: Invalid workflow definition
        '429':
          description: Rate limit exceeded
        '500':
          description: Internal server error
  /workflows/{execution_id}:
    get:
      summary: Get Workflow Result
      description: Get workflow execution result and outputs
      tags:
        - Workflows
      parameters:
        - name: execution_id
          in: path
          required: true
          schema:
            type: string
          description: Workflow execution ID
      responses:
        '200':
          description: Workflow result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResult'
        '404':
          description: Workflow not found
  /workflows/:
    get:
      summary: List Workflow Executions
      description: List user's workflow executions with filtering
      tags:
        - Workflows
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/WorkflowStatus'
          description: Filter by status
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Offset for pagination
      responses:
        '200':
          description: List of workflow executions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowResult'
  /workflows/{execution_id}/cancel:
    post:
      summary: Cancel Workflow
      description: Cancel a running or pending workflow
      tags:
        - Workflows
      parameters:
        - name: execution_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workflow cancelled
        '400':
          description: Cannot cancel workflow
        '404':
          description: Workflow not found
  /workflows/{execution_id}/status:
    get:
      summary: Get Workflow Status
      description: Get real-time workflow status and progress
      tags:
        - Workflows
      parameters:
        - name: execution_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workflow status and progress
          content:
            application/json:
              schema:
                type: object
                properties:
                  execution_id:
                    type: string
                  status:
                    $ref: '#/components/schemas/WorkflowStatus'
                  created_at:
                    type: string
                    format: date-time
                  started_at:
                    type: string
                    format: date-time
                    nullable: true
                  queue_position:
                    type: integer
                    nullable: true
                  progress:
                    type: object
                    nullable: true
                  estimated_completion:
                    type: string
                    format: date-time
                    nullable: true
  /workflows/{execution_id}/logs:
    get:
      summary: Get Workflow Logs
      description: Get workflow execution logs
      tags:
        - Workflows
      parameters:
        - name: execution_id
          in: path
          required: true
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
            enum: [DEBUG, INFO, WARNING, ERROR]
          description: Filter by log level
        - name: tail
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          description: Number of recent logs
      responses:
        '200':
          description: Workflow logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  execution_id:
                    type: string
                  logs:
                    type: array
                    items:
                      type: object
                  total_count:
                    type: integer
                  filtered:
                    type: boolean
  /workflows/{execution_id}/retry:
    post:
      summary: Retry Failed Workflow
      description: Retry a failed workflow execution
      tags:
        - Workflows
      parameters:
        - name: execution_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Retry successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  execution_id:
                    type: string
                  status:
                    $ref: '#/components/schemas/WorkflowStatus'
                  queue_position:
                    type: integer
                  retried_from:
                    type: string

  # Model Management endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    WorkflowStatus:
      type: string
      enum:
        - pending
        - running
        - completed
        - failed
        - cancelled
      example: "pending"
    Priority:
      type: string
      enum:
        - low
        - normal
        - high
      example: "normal"
    WorkflowNodeInput:
      type: object
      required:
        - name
        - type
        - value
      properties:
        name:
          type: string
          example: "seed"
        type:
          type: string
          example: "INT"
        value:
          oneOf:
            - type: string
            - type: number
            - type: boolean
            - type: array
            - type: object
          example: 12345
        required:
          type: boolean
          default: true
    WorkflowNode:
      type: object
      required:
        - id
        - class_type
      properties:
        id:
          type: string
          example: "1"
        class_type:
          type: string
          example: "KSampler"
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowNodeInput'
          default: []
        outputs:
          type: array
          items:
            type: string
          default: []
    WorkflowDefinition:
      type: object
      required:
        - nodes
      properties:
        nodes:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/WorkflowNode'
          example:
            "1":
              id: "1"
              class_type: "CheckpointLoaderSimple"
              inputs:
                - name: "ckpt_name"
                  type: "STRING"
                  value: "v1-5-pruned-emaonly.ckpt"
                  required: true
        metadata:
          type: object
          additionalProperties: true
          default: {}
    WorkflowExecutionRequest:
      type: object
      required:
        - workflow
      properties:
        workflow:
          $ref: '#/components/schemas/WorkflowDefinition'
        priority:
          $ref: '#/components/schemas/Priority'
        webhook_url:
          type: string
          format: uri
          nullable: true
          example: "https://your-app.com/webhook"
        metadata:
          type: object
          additionalProperties: true
          default: {}
        timeout_minutes:
          type: integer
          minimum: 1
          maximum: 120
          default: 30
    WorkflowExecutionResponse:
      type: object
      required:
        - execution_id
        - status
        - created_at
      properties:
        execution_id:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        status:
          $ref: '#/components/schemas/WorkflowStatus'
        created_at:
          type: string
          format: date-time
        estimated_duration:
          type: integer
          nullable: true
          description: "Estimated duration in seconds"
        queue_position:
          type: integer
          nullable: true
          example: 3
    WorkflowResult:
      type: object
      required:
        - execution_id
        - status
      properties:
        execution_id:
          type: string
        status:
          $ref: '#/components/schemas/WorkflowStatus'
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        duration_seconds:
          type: number
          format: float
          nullable: true
        outputs:
          type: object
          additionalProperties: true
          nullable: true
        error:
          type: string
          nullable: true
        logs:
          type: array
          items:
            type: object
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true

    # Model schemas
...